requirements.txt
Django>=5.0,<6
psycopg[binary]>=3.1,<4
strawberry-graphql>=0.220,<1
strawberry-graphql-django>=0.10,<1
python-dotenv>=1.0,<2
PyJWT>=2.8,<3
openpyxl>=3.1,<4
django-cors-headers>=4.3,<5
reportlab>=4.0,<5
